+++ src/main/java/oram/utils/ORAMUtils.java
@@ -11,7 +11,7 @@ import java.util.Collections;
 import java.util.List;
 
 public class ORAMUtils {
-	public static final double DUMMY_VERSION = 0;
+	public static final int DUMMY_VERSION = 0;
 	public static final byte DUMMY_PATH = -1;
 	public static final int DUMMY_ADDRESS = -1;
 	public static final byte[] DUMMY_BLOCK = new byte[0];
